requirements.txt
Django>=4.2,<5.0
Pillow>=9.0.0
python-decouple>=3.6
djangorestframework>=3.16.0
django-cors-headers>=4.8.0
